watcher/requirements.txt
imagehash
Pillow
PyYAML
